Fetch requests back into sw.js fixes global worker errors
